# BUG: Better tree pruning for complex queries for improved performance