Closes #859. There's still some redundant work happening, but further tweaking can happen in a follow up PR. From the example in #889, we have:

Before this PR:

```
In [2]: %time ibis.impala.compile(src_table)
CPU times: user 2.08 s, sys: 0 ns, total: 2.08 s
Wall time: 2.08 s
```

after:

```
In [2]: %timeit _ = ibis.impala.compile(src_table)
10 loops, best of 3: 57.9 ms per loop
```

This bad compilation performance is pretty vicious, so I think it's time for a 0.9 release. Sound good @cpcloud?