@cpcloud 2016-10-30
@wesm Yep, agree with releasing.